requirements.txt
fastapi[standard]
line-bot-sdk==3.14.2
python-dotenv==1.0.1